refactor(bubble-ui): URL 文法を universe@<node> に統一し、root URL をパスへ移す
これまで世界線の現在ノードは、
  - root はブラウザのハッシュ #wl=<node>
  - ネストは親バブル url の query 文字列 universe?wl=<node>
と書いていた。この commit で両者を「<name>@<node>」という単一文法に統一する：
  - root はブラウザのパス /universe@<node>
  - ネストは親バブル url universe@<node>

理由：

- 「wl」は world line で一般語すぎる。重要なのは「universe を このスナップショットで
  見ている」という関係で、git HEAD@{N} / docker image@digest / npm pkg@version 等と
  同じ「at this snapshot」イデオムに乗れる
- ? は「複数 key を扱う一般装置」の構文だが、付けたいのは node 1個だけなので
  <name>@<val> の方が cardinality に合う
- # は本来「同一ドキュメント内のアンカー」を指す記号で、SPA 状態の保管庫として使うのは
  記号本来の意味とズレる。Next.js なら パスで素直に書ける
- 結果として root と nest が完全に同じ形 universe@<node> になり、再帰的入れ子の
  対称性が url 表現でも保たれる

変更点：

- wl-url.ts (WL_URL_KEY = "wl") を削除し、snapshot-url.ts に置き換え
  - UNIVERSE_URL_BASE = "universe"
  - AT_PREFIX = "@"
  - buildUniverseUrl(node) → "universe@<node>"
  - readUniverseAt(url) → "universe@<node>" から node 抽出
- useUniverseWorldLine.ts: ネスト側のヘルパーを snapshot-url.ts に委譲
- bubbleRoutes.tsx: universe ルートの regex を /^universe(@[^/?#&]+)?$/ に
- useBubbleArrangementWorldLine.ts: root のブラウザ url を hash から path へ書き換え
  - parseNodeFromUrl は location.pathname を見る
  - pushState/replaceState は "/universe@<node>" を書き込む
- app/page.tsx を app/[[...slug]]/page.tsx に移動（optional catch-all で / と
  /universe@<node> を同じ SPA に流す。/world-line 等の specific page は静的優先で先勝ち）
- docs/recursive-universe.md を新文法に追従。「なぜ パス か」「なぜ @ か」を設計判断
  セクションに追加

将来 users/<id>@<node> 等を他オブジェクトに付けたくなったときも、同じ文法で揃う
（文法予約。コードはまだ無い）。
